+++ RxGRDB/Result.swift
@@ -1,49 +1,7 @@
-//
-//  Result.swift
-//
-//  Copyright (c) 2014-2016 Alamofire Software Foundation (http://alamofire.org/)
-//
-//  Permission is hereby granted, free of charge, to any person obtaining a copy
-//  of this software and associated documentation files (the "Software"), to deal
-//  in the Software without restriction, including without limitation the rights
-//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-//  copies of the Software, and to permit persons to whom the Software is
-//  furnished to do so, subject to the following conditions:
-//
-//  The above copyright notice and this permission notice shall be included in
-//  all copies or substantial portions of the Software.
-//
-//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-//  THE SOFTWARE.
-//
-
 enum Result<Value> {
     case success(Value)
     case failure(Error)
     
-    /// Creates a `Result` instance from the result of a closure.
-    ///
-    /// A failure result is created when the closure throws, and a success result is created when the closure
-    /// succeeds without throwing an error.
-    ///
-    ///     func someString() throws -> String { ... }
-    ///
-    ///     let result = Result(value: {
-    ///         return try someString()
-    ///     })
-    ///
-    ///     // The type of result is Result<String>
-    ///
-    /// The trailing closure syntax is also supported:
-    ///
-    ///     let result = Result { try someString() }
-    ///
-    /// - parameter value: The closure to execute and create the result for.
     public init(value: () throws -> Value) {
         do {
             self = try .success(value())
@@ -52,25 +10,6 @@ enum Result<Value> {
         }
     }
     
-    /// Evaluates the given closure when this `Result` is a success, passing the
-    /// unwrapped value as a parameter.
-    ///
-    /// Use the `map` method with a closure that does not throw. For example:
-    ///
-    ///     let possibleData: Result<Data> = .success(Data())
-    ///     let possibleInt = possibleData.map { $0.count }
-    ///     try print(possibleInt.unwrap())
-    ///     // Prints "0"
-    ///
-    ///     let noData: Result<Data> = .failure(error)
-    ///     let noInt = noData.map { $0.count }
-    ///     try print(noInt.unwrap())
-    ///     // Throws error
-    ///
-    /// - parameter transform: A closure that takes the success value of
-    ///   the instance.
-    /// - returns: A `Result` containing the result of the given closure. If
-    ///   this instance is a failure, returns the same failure.
     func map<T>(_ transform: (Value) -> T) -> Result<T> {
         switch self {
         case .success(let value):