+++ CHANGELOG.md
@@ -1,6 +1,28 @@
 Release Notes
 =============
 
+## Next Version
+
+**New**
+
+- The new `distinctUntilChanged` parameter has RxGRDB avoid notifying consecutive identical values.
+
+    ```swift
+    request.rx.fetchAll(in: dbQueue, distinctUntilChanged: true)...
+    ```
+
+- Tracking of requests that fetch an array of optional values:
+    
+    ```swift
+    // Email column may be NULL:
+    let request = Person.select(email).bound(to: Optional<String>.self)
+    request.rx.fetchAll(in: dbQueue)
+        .subscribe(onNext: { emails: [String?] in
+            ...
+        })
+    ```
+
+
 ## v0.2.0
 
 Released May 17, 2017